Implemented game loop logic

+++ src/ECS2/Time.h
@@ -14,7 +14,7 @@
 class Time {
 public:
     // Current time in milliseconds
-    static int Now() {
+    static long Now() {
         return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now().time_since_epoch()).count();
     }
 };
